Refactor class_views/Categorie.py, class_views/Comment.py, class_views/Product.py, models.py, templates, views.py, urls.py, and views.py for better code organization and readability

+++ plataforma/models.py
@@ -35,7 +35,7 @@ class Categorie(models.Model):
 class Comment(models.Model):
     title = models.CharField(max_length=200)
     description = models.TextField()
-    stars = models.FloatField()
+    stars = models.IntegerField()
     image = models.ImageField(upload_to='comments/images/', )
 
     def __str__(self):
